build<pom.xml>: hibernate validator

+++ src/main/java/com/simpletodolist/simpletodolist/service/TaskService.java
@@ -3,6 +3,7 @@ package com.simpletodolist.simpletodolist.service;
 import com.simpletodolist.simpletodolist.dto.TaskDto;
 import com.simpletodolist.simpletodolist.dto.TaskWithDateDto;
 import com.simpletodolist.simpletodolist.entity.Task;
+import com.simpletodolist.simpletodolist.exception.NotFoundException;
 import com.simpletodolist.simpletodolist.repository.TaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -44,6 +45,10 @@ public class TaskService {
     }
 
     public String deleteTask(Long id) {
+        Optional<Task> taskOptional = taskRepository.findById(id);
+        if (taskOptional.isEmpty()) {
+            throw new NotFoundException("Id " + id + " does not match any task");
+        }
         taskRepository.deleteById(id);
         return "Tarefa com o id " + id + " apagada com sucesso";
     }
